fix: fix CI/CD issues and all type/test errors
- fix(ci): Trivy action version pinned to v0.35.0
- fix(ci): add security job to all-passed dependency
- refactor(backend): move evaluation module from tests/ to src/
- fix(backend): fix all 168 MyPy type errors
- test(frontend): fix all 35 failing frontend tests

Version: 2.6.6

+++ backend/src/raghub_mcp/rerank_engine/scorers/vector_scorer.py
@@ -205,7 +205,7 @@ class VectorScorer(BaseScorer):
             Similarity in range (0, 1], where 1 means identical.
         """
         distance = np.linalg.norm(vec1 - vec2)
-        return 1.0 / (1.0 + distance)
+        return float(1.0 / (1.0 + float(distance)))
 
     def get_config(self) -> dict[str, Any]:
         """Get scorer configuration for logging/debugging.
